`feat`: auth refactor + get_current_user
