+++ backend/src/security/__init__.py
@@ -1,8 +1,9 @@
+from .auth import create_access_token, get_current_user
 from .hashing import get_password_hash, verify_password
-from .jwt import create_access_token
 
 __all__ = [
     'create_access_token',
+    'get_current_user',
     'get_password_hash',
     'verify_password',
 ]
