--- backend/src/security/jwt.py
@@ -1,22 +1,0 @@
-from datetime import datetime, timedelta
-
-from jwt import encode
-from zoneinfo import ZoneInfo
-
-from src.settings import Settings
-
-SECRET_KEY = Settings().SECRET_KEY
-ALGORITHM = Settings().ALGORITHM
-ACCESS_TOKEN_EXPIRE_MINUTES = Settings().ACCESS_TOKEN_EXPIRE_MINUTES
-
-
-def create_access_token(data: dict) -> str:
-    to_encode = data.copy()
-
-    expire = datetime.now(tz=ZoneInfo('UTC')) + timedelta(
-        minutes=ACCESS_TOKEN_EXPIRE_MINUTES
-    )
-    to_encode.update({'exp': expire})
-
-    encoded_jwt = encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-    return encoded_jwt
